+++ Assets/Plugins/Mediabox/Source/GameKit/Pause/PauseHandle.cs
@@ -1,0 +1,4 @@
+namespace Mediabox.GameKit.Pause
+{
+    public class PauseHandle{}
+}